DataFrame,Plotting,PlottingGL,Volume,WebBrowser: renamed file of DataFrameColumnProperty
