+++ modules/dataframe/src/properties/dataframecolumnproperty.cpp
@@ -27,7 +27,7 @@
  *
  *********************************************************************************/
 
-#include <inviwo/dataframe/properties/dataframeproperty.h>
+#include <inviwo/dataframe/properties/dataframecolumnproperty.h>
 
 namespace inviwo {
