+++ modules/plotting/include/modules/plotting/processors/dataframecolumntocolorvector.h
@@ -37,7 +37,7 @@
 #include <inviwo/core/ports/dataoutport.h>
 #include <inviwo/core/properties/transferfunctionproperty.h>
 #include <inviwo/dataframe/datastructures/dataframe.h>
-#include <inviwo/dataframe/properties/dataframeproperty.h>
+#include <inviwo/dataframe/properties/dataframecolumnproperty.h>
 
 namespace inviwo {
