+++ modules/plottinggl/include/modules/plottinggl/processors/scatterplotmatrixprocessor.h
@@ -40,7 +40,7 @@
 
 #include <inviwo/dataframe/datastructures/dataframe.h>
 #include <modules/plottinggl/plotters/scatterplotgl.h>
-#include <inviwo/dataframe/properties/dataframeproperty.h>
+#include <inviwo/dataframe/properties/dataframecolumnproperty.h>
 #include <modules/opengl/texture/texture2d.h>
 #include <modules/opengl/rendering/texturequadrenderer.h>
 #include <modules/fontrendering/textrenderer.h>
