+++ modules/plottinggl/include/modules/plottinggl/processors/scatterplotprocessor.h
@@ -34,7 +34,7 @@
 #include <inviwo/core/ports/imageport.h>
 #include <inviwo/dataframe/datastructures/dataframe.h>
 #include <modules/plottinggl/plotters/scatterplotgl.h>
-#include <inviwo/dataframe/properties/dataframeproperty.h>
+#include <inviwo/dataframe/properties/dataframecolumnproperty.h>
 #include <modules/brushingandlinking/ports/brushingandlinkingports.h>
 
 namespace inviwo {
